Changement int en long pour les timestamp. Correction bug dans le cache (lastGlobalUpdate non mis à jour).
git-svn-id: https://svn.itinerennes.fr/android/trunk@159 3fbbf0ad-bdca-4ea9-bcf4-e3b57373991e

+++ ItineRennes/src/fr/itinerennes/business/service/LineIconService.java
@@ -46,7 +46,7 @@ public class LineIconService extends AbstractService {
     private final CacheProvider<LineIcon> iconCache;
 
     /** The last time all the cache was updated (in seconds). */
-    private int lastGlobalUpdate = 0;
+    private long lastGlobalUpdate = 0;
 
     /**
      * Creates a line transport icon service.
